requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iso8601==1.1.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
pypika-tortoise==0.1.6
pyscopg2==66.0.2
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.22.0
